Completed 4.23

+++ tests/blog_api.test.js
@@ -2,6 +2,7 @@ const supertest = require('supertest')
 const mongoose = require('mongoose')
 const app = require('../app')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 const { initial } = require('lodash')
 const api = supertest(app)
 
@@ -19,14 +20,32 @@ const initalBlogs = [
         url: "http://www.u.arizona.edu/~rubinson/copyright_violations/Go_To_Considered_Harmful.html",
         likes: 5
         }
+        
 ]
+
 
 beforeEach(async () => {
     await Blog.deleteMany({})
+    await User.deleteMany({})
     let newBlog = new Blog(initalBlogs[0])
+    let newBlog2 = new Blog(initalBlogs[1])
+    let newUser = {
+        username: "lol123",
+        password: "12345",
+        name: "Jordan"
+    }
+    const result = await api.post('/api/users').send(newUser)
+
+    console.log("result is", result._body.savedUser.id)
+
+    newBlog.user = result._body.savedUser.id
+    newBlog2.user = result._body.savedUser.id
+
     await newBlog.save()
-    newBlog = new Blog(initalBlogs[1])
-    await newBlog.save()
+    await newBlog2.save()
+    
+
+
 })
 
 describe('blog tests', () => {
@@ -51,11 +70,15 @@ describe('blog tests', () => {
             url: "https://fregrferfref.com/",
             likes: 7
             }
-        const result = await api.post('/api/blogs').send(newBlog).expect(201)
+        
+        const loginRes = await api.post('/api/login').send({username:'lol123', password: '12345'})
+        console.log("loginRes is", loginRes.body)
+
+        const result = await api.post('/api/blogs').send(newBlog).set({Authorization: `bearer ${loginRes.body.token}`}).expect(201)
         const secondResult = await api.get('/api/blogs')
 
         console.log("Second result is", secondResult._body)
-        const filteredArr = secondResult._body.map(a => a.title);
+        const filteredArr = secondResult._body.map(a => a.title)
         expect(secondResult._body).toHaveLength(initalBlogs.length + 1)
         expect(filteredArr).toContain('Node patterns')
         
@@ -72,7 +95,11 @@ describe('blog tests', () => {
                 newBlog.likes = 0;
             }
 
-            const blogs = await api.post('/api/blogs').send(newBlog)
+            const loginRes = await api.post('/api/login').send({username:'lol123', password: '12345'})
+            console.log("loginRes is", loginRes.body)
+
+            const blogs = await api.post('/api/blogs').send(newBlog).set({Authorization: `bearer ${loginRes.body.token}`}).expect(201)
+            // const blogs = await api.post('/api/blogs').send(newBlog)
             const blogList = await api.get('/api/blogs')
 
             const blogBody = blogList._body
@@ -87,24 +114,58 @@ describe('blog tests', () => {
             author: "Cool Chan",
             likes:250
             }
-        const result = await api.post('/api/blogs').send(newBlog).expect(400)
+        
+        const loginRes = await api.post('/api/login').send({username:'lol123', password: '12345'})
+        await api.post('/api/blogs').send(newBlog).set({Authorization: `bearer ${loginRes.body.token}`}).expect(400)
+        
+        // const result = await api.post('/api/blogs').send(newBlog).expect(400)
     })
 
     test('test if deleting blog works', async () => {
-        const id = '6287e0b51b578336fd02a996'
-        const result = await api.delete(`/api/blogs/${id}`).expect(JSON.stringify("Successfully deleted blog"))
+        const loginRes = await api.post('/api/login').send({username:'lol123', password: '12345'})
+        
+        const blogToDelete = await Blog.find({
+            title: "React patterns",
+            author: "Michael Chan",
+            url: "https://reactpatterns.com/",
+            likes: 7
+            })
+
+        console.log("blog to delete is", blogToDelete)
+        const id = blogToDelete[0]._id
+        
+        const result = await api.delete(`/api/blogs/${id}`).set({Authorization: `bearer ${loginRes.body.token}`}).expect(JSON.stringify("Successfully deleted blog"))
     })
 
     test('test to update blog', async () => {
         const blogList = await api.get('/api/blogs')
 
         console.log("BLOGLIST IS", blogList._body)
-        const id = '6289724d4ac29adc0e306841'
+        const blogToUpdate = await Blog.find({
+            title: "React patterns",
+            author: "Michael Chan",
+            url: "https://reactpatterns.com/",
+            likes: 7
+            })
+        const id = blogToUpdate[0]._id
+
         const newBlog = {
             likes: 15
         }
 
         const updated = await api.put(`/api/blogs/${id}`).send(newBlog).expect(JSON.stringify("Successfully updated blog"))
+    })
+
+    test('send 404 error if token is not provided', async() => {
+        const newBlog = {
+            title: "Node patterns",
+            author: "Cool Chan",
+            url: "https://fregrferfref.com/",
+            likes: 7
+            }
+
+        const loginRes = await api.post('/api/login').send({username:'lol123', password: '12345'})
+        await api.post('/api/blogs').send(newBlog).expect(401)
     })
 })
 
@@ -116,3 +177,9 @@ afterAll(() => {
 //npm test -- tests/blog_api.test.js --silent=false
 
 //npm test -- tests/blog_api.test.js --verbose false
+
+//npm test -- -t "adds new blog"
+
+//npm test -- -t "send 404 error if token is not provided"
+
+//npm test -- -t "test if deleting blog works"
